+++ NEWS.md
@@ -35,6 +35,8 @@ polmineR 0.7.11.9001
 * The `enrich()`-method for objects of class `kwic` now includes a new argument `extra` that will add extra tokens to the left of the windows for concordances so that qualitative inspections for query hits can work with more context.
 * The `label()`-method has been renamed as `edit()`-method now.
 * Indexing the concordances of a `kwic` object did not reduce the `cpos` table concurringly, corrected.
+* The table in the `table` slot of `kwic` objects has been turned into a `data.table`.
+* A new `kwic_bundle`-class has been introduced, a list of `kwic` objects can be turned into this new class using `as.bundle`.
 
 
 polmineR 0.7.11